refactor(connector): extract IcebergCommon config
