+++ src/connector/src/sink/iceberg/mod.rs
@@ -12,31 +12,20 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-mod jni_catalog;
-mod mock_catalog;
 mod prometheus;
-mod storage_catalog;
 
 use std::collections::{BTreeMap, HashMap};
 use std::fmt::Debug;
 use std::num::NonZeroU64;
 use std::ops::Deref;
-use std::sync::Arc;
 
 use anyhow::{anyhow, Context};
 use arrow_schema_iceberg::{
     DataType as ArrowDataType, Field as ArrowField, Fields, Schema as ArrowSchema, SchemaRef,
 };
 use async_trait::async_trait;
-use iceberg::io::{S3_ACCESS_KEY_ID, S3_ENDPOINT, S3_REGION, S3_SECRET_ACCESS_KEY};
-use iceberg::spec::TableMetadata;
-use iceberg::table::Table as TableV2;
-use iceberg::{Catalog as CatalogV2, NamespaceIdent, TableCreation, TableIdent};
-use iceberg_catalog_glue::{AWS_ACCESS_KEY_ID, AWS_REGION_NAME, AWS_SECRET_ACCESS_KEY};
-use icelake::catalog::{
-    load_catalog, load_iceberg_base_catalog_config, BaseCatalogConfig, CatalogRef, CATALOG_NAME,
-    CATALOG_TYPE,
-};
+use iceberg::{NamespaceIdent, TableCreation};
+use icelake::catalog::CatalogRef;
 use icelake::io_v2::input_wrapper::{DeltaWriter, RecordBatchWriter};
 use icelake::io_v2::prometheus::{PrometheusWriterBuilder, WriterMetrics};
 use icelake::io_v2::{
@@ -44,7 +33,7 @@ use icelake::io_v2::{
 };
 use icelake::transaction::Transaction;
 use icelake::types::{data_file_from_json, data_file_to_json, Any, DataFile};
-use icelake::{Table, TableIdentifier};
+use icelake::Table;
 use itertools::Itertools;
 use risingwave_common::array::arrow::{IcebergArrowConvert, IcebergCreateTableArrowConvert};
 use risingwave_common::array::{Op, StreamChunk};
@@ -56,12 +45,9 @@ use risingwave_pb::connector_service::sink_metadata::SerializedMetadata;
 use risingwave_pb::connector_service::SinkMetadata;
 use serde_derive::Deserialize;
 use serde_with::{serde_as, DisplayFromStr};
-use storage_catalog::StorageCatalogConfig;
 use thiserror_ext::AsReport;
-use url::Url;
 use with_options::WithOptions;
 
-use self::mock_catalog::MockCatalog;
 use self::prometheus::monitored_base_file_writer::MonitoredBaseFileWriterBuilder;
 use self::prometheus::monitored_position_delete_writer::MonitoredPositionDeleteWriterBuilder;
 use super::decouple_checkpoint_log_sink::{
@@ -70,7 +56,7 @@ use super::decouple_checkpoint_log_sink::{
 use super::{
     Sink, SinkError, SinkWriterParam, SINK_TYPE_APPEND_ONLY, SINK_TYPE_OPTION, SINK_TYPE_UPSERT,
 };
-use crate::error::ConnectorResult;
+use crate::connector_common::IcebergCommon;
 use crate::sink::coordinate::CoordinatedSinkWriter;
 use crate::sink::writer::SinkWriter;
 use crate::sink::{Result, SinkCommitCoordinator, SinkParam};
@@ -79,53 +65,18 @@ use crate::{
     deserialize_optional_string_seq_from_string,
 };
 
-/// This iceberg sink is WIP. When it ready, we will change this name to "iceberg".
 pub const ICEBERG_SINK: &str = "iceberg";
 
 #[serde_as]
-#[derive(Debug, Clone, PartialEq, Eq, Deserialize, WithOptions, Default)]
+#[derive(Debug, Clone, PartialEq, Eq, Deserialize, WithOptions)]
 pub struct IcebergConfig {
-    pub connector: String, // Avoid deny unknown field. Must be "iceberg"
-
     pub r#type: String, // accept "append-only" or "upsert"
 
     #[serde(default, deserialize_with = "deserialize_bool_from_string")]
     pub force_append_only: bool,
 
-    #[serde(rename = "table.name")]
-    pub table_name: String, // Full name of table, must include schema name
-
-    #[serde(rename = "database.name")]
-    pub database_name: Option<String>,
-    // Database name of table
-
-    // Catalog name, can be omitted for storage catalog, but
-    // must be set for other catalogs.
-    #[serde(rename = "catalog.name")]
-    pub catalog_name: Option<String>,
-
-    // Catalog type supported by iceberg, such as "storage", "rest".
-    // If not set, we use "storage" as default.
-    #[serde(rename = "catalog.type")]
-    pub catalog_type: Option<String>,
-
-    #[serde(rename = "warehouse.path")]
-    pub path: String, // Path of iceberg warehouse, only applicable in storage catalog.
-
-    #[serde(rename = "catalog.uri")]
-    pub uri: Option<String>, // URI of iceberg catalog, only applicable in rest catalog.
-
-    #[serde(rename = "s3.region")]
-    pub region: Option<String>,
-
-    #[serde(rename = "s3.endpoint")]
-    pub endpoint: Option<String>,
-
-    #[serde(rename = "s3.access.key")]
-    pub access_key: String,
-
-    #[serde(rename = "s3.secret.key")]
-    pub secret_key: String,
+    #[serde(flatten)]
+    pub common: IcebergCommon,
 
     #[serde(
         rename = "s3.path.style.access",
@@ -185,7 +136,9 @@ impl IcebergConfig {
             }
         }
 
-        if config.catalog_name.is_none() && config.catalog_type.as_deref() != Some("storage") {
+        if config.common.catalog_name.is_none()
+            && config.common.catalog_type.as_deref() != Some("storage")
+        {
             return Err(SinkError::Config(anyhow!(
                 "catalog.name must be set for non-storage catalog"
             )));
@@ -210,469 +163,6 @@ impl IcebergConfig {
         }
 
         Ok(config)
-    }
-
-    pub fn catalog_type(&self) -> &str {
-        self.catalog_type.as_deref().unwrap_or("storage")
-    }
-
-    fn catalog_name(&self) -> String {
-        self.catalog_name
-            .as_ref()
-            .map(|s| s.to_string())
-            .unwrap_or_else(|| "risingwave".to_string())
-    }
-
-    fn full_table_name(&self) -> Result<TableIdentifier> {
-        let ret = if let Some(database_name) = &self.database_name {
-            TableIdentifier::new(vec![database_name, &self.table_name])
-        } else {
-            TableIdentifier::new(vec![&self.table_name])
-        };
-
-        ret.context("Failed to create table identifier")
-            .map_err(|e| SinkError::Iceberg(anyhow!(e)))
-    }
-
-    fn build_iceberg_configs(&self) -> Result<HashMap<String, String>> {
-        let mut iceberg_configs = HashMap::new();
-
-        let catalog_type = self.catalog_type().to_string();
-
-        iceberg_configs.insert(CATALOG_TYPE.to_string(), catalog_type.clone());
-        iceberg_configs.insert(CATALOG_NAME.to_string(), self.catalog_name());
-
-        match catalog_type.as_str() {
-            "storage" => {
-                iceberg_configs.insert(
-                    format!("iceberg.catalog.{}.warehouse", self.catalog_name()),
-                    self.path.clone(),
-                );
-            }
-            "rest" => {
-                let uri = self.uri.clone().ok_or_else(|| {
-                    SinkError::Iceberg(anyhow!("`catalog.uri` must be set in rest catalog"))
-                })?;
-                iceberg_configs.insert(format!("iceberg.catalog.{}.uri", self.catalog_name()), uri);
-            }
-            _ => {
-                return Err(SinkError::Iceberg(anyhow!(
-                    "Unsupported catalog type: {}, only support `storage` and `rest`",
-                    catalog_type
-                )));
-            }
-        }
-
-        if let Some(region) = &self.region {
-            iceberg_configs.insert(
-                "iceberg.table.io.region".to_string(),
-                region.clone().to_string(),
-            );
-        }
-
-        if let Some(endpoint) = &self.endpoint {
-            iceberg_configs.insert(
-                "iceberg.table.io.endpoint".to_string(),
-                endpoint.clone().to_string(),
-            );
-        }
-
-        iceberg_configs.insert(
-            "iceberg.table.io.access_key_id".to_string(),
-            self.access_key.clone().to_string(),
-        );
-        iceberg_configs.insert(
-            "iceberg.table.io.secret_access_key".to_string(),
-            self.secret_key.clone().to_string(),
-        );
-        if let Some(path_style_access) = self.path_style_access {
-            iceberg_configs.insert(
-                "iceberg.table.io.enable_virtual_host_style".to_string(),
-                (!path_style_access).to_string(),
-            );
-        }
-
-        let (bucket, root) = {
-            let url = Url::parse(&self.path).map_err(|e| SinkError::Iceberg(anyhow!(e)))?;
-            let bucket = url
-                .host_str()
-                .ok_or_else(|| {
-                    SinkError::Iceberg(anyhow!("Invalid s3 path: {}, bucket is missing", self.path))
-                })?
-                .to_string();
-            let root = url.path().trim_start_matches('/').to_string();
-            (bucket, root)
-        };
-
-        iceberg_configs.insert("iceberg.table.io.bucket".to_string(), bucket);
-
-        // Only storage catalog should set this.
-        if catalog_type == "storage" {
-            iceberg_configs.insert("iceberg.table.io.root".to_string(), root);
-        }
-        // #TODO
-        // Support load config file
-        iceberg_configs.insert(
-            "iceberg.table.io.disable_config_load".to_string(),
-            "true".to_string(),
-        );
-
-        Ok(iceberg_configs)
-    }
-
-    fn build_jni_catalog_configs(&self) -> Result<(BaseCatalogConfig, HashMap<String, String>)> {
-        let mut iceberg_configs = HashMap::new();
-
-        let base_catalog_config = {
-            let catalog_type = self.catalog_type().to_string();
-
-            iceberg_configs.insert(CATALOG_TYPE.to_string(), catalog_type.clone());
-            iceberg_configs.insert(CATALOG_NAME.to_string(), self.catalog_name());
-
-            if let Some(region) = &self.region {
-                // icelake
-                iceberg_configs.insert(
-                    "iceberg.table.io.region".to_string(),
-                    region.clone().to_string(),
-                );
-                // iceberg-rust
-                iceberg_configs.insert(
-                    ("iceberg.table.io.".to_string() + S3_REGION).to_string(),
-                    region.clone().to_string(),
-                );
-            }
-
-            if let Some(endpoint) = &self.endpoint {
-                iceberg_configs.insert(
-                    "iceberg.table.io.endpoint".to_string(),
-                    endpoint.clone().to_string(),
-                );
-
-                // iceberg-rust
-                iceberg_configs.insert(
-                    ("iceberg.table.io.".to_string() + S3_ENDPOINT).to_string(),
-                    endpoint.clone().to_string(),
-                );
-            }
-
-            // icelake
-            iceberg_configs.insert(
-                "iceberg.table.io.access_key_id".to_string(),
-                self.access_key.clone().to_string(),
-            );
-            iceberg_configs.insert(
-                "iceberg.table.io.secret_access_key".to_string(),
-                self.secret_key.clone().to_string(),
-            );
-
-            // iceberg-rust
-            iceberg_configs.insert(
-                ("iceberg.table.io.".to_string() + S3_ACCESS_KEY_ID).to_string(),
-                self.access_key.clone().to_string(),
-            );
-            iceberg_configs.insert(
-                ("iceberg.table.io.".to_string() + S3_SECRET_ACCESS_KEY).to_string(),
-                self.secret_key.clone().to_string(),
-            );
-
-            let (bucket, _) = {
-                let url = Url::parse(&self.path).map_err(|e| SinkError::Iceberg(anyhow!(e)))?;
-                let bucket = url
-                    .host_str()
-                    .ok_or_else(|| {
-                        SinkError::Iceberg(anyhow!(
-                            "Invalid s3 path: {}, bucket is missing",
-                            self.path
-                        ))
-                    })?
-                    .to_string();
-                let root = url.path().trim_start_matches('/').to_string();
-                (bucket, root)
-            };
-
-            iceberg_configs.insert("iceberg.table.io.bucket".to_string(), bucket);
-            // #TODO
-            // Support load config file
-            iceberg_configs.insert(
-                "iceberg.table.io.disable_config_load".to_string(),
-                "true".to_string(),
-            );
-
-            load_iceberg_base_catalog_config(&iceberg_configs)?
-        };
-
-        // Prepare jni configs, for details please see https://iceberg.apache.org/docs/latest/aws/
-        let mut java_catalog_configs = HashMap::new();
-        {
-            if let Some(uri) = self.uri.as_deref() {
-                java_catalog_configs.insert("uri".to_string(), uri.to_string());
-            }
-
-            java_catalog_configs.insert("warehouse".to_string(), self.path.clone());
-            java_catalog_configs.extend(self.java_catalog_props.clone());
-
-            // Currently we only support s3, so let's set it to s3
-            java_catalog_configs.insert(
-                "io-impl".to_string(),
-                "org.apache.iceberg.aws.s3.S3FileIO".to_string(),
-            );
-
-            if let Some(endpoint) = &self.endpoint {
-                java_catalog_configs
-                    .insert("s3.endpoint".to_string(), endpoint.clone().to_string());
-            }
-
-            java_catalog_configs.insert(
-                "s3.access-key-id".to_string(),
-                self.access_key.clone().to_string(),
-            );
-            java_catalog_configs.insert(
-                "s3.secret-access-key".to_string(),
-                self.secret_key.clone().to_string(),
-            );
-
-            if let Some(path_style_access) = self.path_style_access {
-                java_catalog_configs.insert(
-                    "s3.path-style-access".to_string(),
-                    path_style_access.to_string(),
-                );
-            }
-            if matches!(self.catalog_type.as_deref(), Some("glue")) {
-                java_catalog_configs.insert(
-                    "client.credentials-provider".to_string(),
-                    "com.risingwave.connector.catalog.GlueCredentialProvider".to_string(),
-                );
-                // Use S3 ak/sk and region as glue ak/sk and region by default.
-                // TODO: use different ak/sk and region for s3 and glue.
-                java_catalog_configs.insert(
-                    "client.credentials-provider.glue.access-key-id".to_string(),
-                    self.access_key.clone().to_string(),
-                );
-                java_catalog_configs.insert(
-                    "client.credentials-provider.glue.secret-access-key".to_string(),
-                    self.secret_key.clone().to_string(),
-                );
-                if let Some(region) = &self.region {
-                    java_catalog_configs
-                        .insert("client.region".to_string(), region.clone().to_string());
-                    java_catalog_configs.insert(
-                        "glue.endpoint".to_string(),
-                        format!("https://glue.{}.amazonaws.com", region),
-                    );
-                }
-            }
-        }
-
-        Ok((base_catalog_config, java_catalog_configs))
-    }
-
-    async fn create_catalog(&self) -> ConnectorResult<CatalogRef> {
-        match self.catalog_type() {
-            "storage" | "rest" => {
-                let iceberg_configs = self.build_iceberg_configs()?;
-                let catalog = load_catalog(&iceberg_configs).await?;
-                Ok(catalog)
-            }
-            catalog_type
-                if catalog_type == "hive" || catalog_type == "jdbc" || catalog_type == "glue" =>
-            {
-                // Create java catalog
-                let (base_catalog_config, java_catalog_props) = self.build_jni_catalog_configs()?;
-                let catalog_impl = match catalog_type {
-                    "hive" => "org.apache.iceberg.hive.HiveCatalog",
-                    "jdbc" => "org.apache.iceberg.jdbc.JdbcCatalog",
-                    "glue" => "org.apache.iceberg.aws.glue.GlueCatalog",
-                    _ => unreachable!(),
-                };
-
-                jni_catalog::JniCatalog::build_catalog(
-                    base_catalog_config,
-                    self.catalog_name(),
-                    catalog_impl,
-                    java_catalog_props,
-                )
-            }
-            "mock" => Ok(Arc::new(MockCatalog {})),
-            _ => {
-                bail!(
-                    "Unsupported catalog type: {}, only support `storage`, `rest`, `hive`, `jdbc`, `glue`",
-                    self.catalog_type()
-                )
-            }
-        }
-    }
-
-    pub async fn load_table(&self) -> ConnectorResult<Table> {
-        let catalog = self
-            .create_catalog()
-            .await
-            .context("Unable to load iceberg catalog")?;
-
-        let table_id = self
-            .full_table_name()
-            .context("Unable to parse table name")?;
-
-        catalog.load_table(&table_id).await.map_err(Into::into)
-    }
-}
-
-impl IcebergConfig {
-    fn full_table_name_v2(&self) -> Result<TableIdent> {
-        let ret = if let Some(database_name) = &self.database_name {
-            TableIdent::from_strs(vec![database_name, &self.table_name])
-        } else {
-            TableIdent::from_strs(vec![&self.table_name])
-        };
-
-        ret.context("Failed to create table identifier")
-            .map_err(|e| SinkError::Iceberg(anyhow!(e)))
-    }
-
-    async fn create_catalog_v2(&self) -> ConnectorResult<Arc<dyn CatalogV2>> {
-        match self.catalog_type() {
-            "storage" => {
-                let config = StorageCatalogConfig::builder()
-                    .warehouse(self.path.clone())
-                    .access_key(self.access_key.clone())
-                    .secret_key(self.secret_key.clone())
-                    .region(self.region.clone())
-                    .endpoint(self.endpoint.clone())
-                    .build();
-                let catalog = storage_catalog::StorageCatalog::new(config)?;
-                Ok(Arc::new(catalog))
-            }
-            "rest" => {
-                let mut iceberg_configs = HashMap::new();
-                if let Some(region) = &self.region {
-                    iceberg_configs.insert(S3_REGION.to_string(), region.clone().to_string());
-                }
-                if let Some(endpoint) = &self.endpoint {
-                    iceberg_configs.insert(S3_ENDPOINT.to_string(), endpoint.clone().to_string());
-                }
-                iceberg_configs.insert(
-                    S3_ACCESS_KEY_ID.to_string(),
-                    self.access_key.clone().to_string(),
-                );
-                iceberg_configs.insert(
-                    S3_SECRET_ACCESS_KEY.to_string(),
-                    self.secret_key.clone().to_string(),
-                );
-                let config = iceberg_catalog_rest::RestCatalogConfig::builder()
-                    .uri(self.uri.clone().ok_or_else(|| {
-                        SinkError::Iceberg(anyhow!("`catalog.uri` must be set in rest catalog"))
-                    })?)
-                    .props(iceberg_configs)
-                    .build();
-                let catalog = iceberg_catalog_rest::RestCatalog::new(config);
-                Ok(Arc::new(catalog))
-            }
-            "glue" => {
-                let mut iceberg_configs = HashMap::new();
-                // glue
-                if let Some(region) = &self.region {
-                    iceberg_configs.insert(AWS_REGION_NAME.to_string(), region.clone().to_string());
-                }
-                iceberg_configs.insert(
-                    AWS_ACCESS_KEY_ID.to_string(),
-                    self.access_key.clone().to_string(),
-                );
-                iceberg_configs.insert(
-                    AWS_SECRET_ACCESS_KEY.to_string(),
-                    self.secret_key.clone().to_string(),
-                );
-                // s3
-                if let Some(region) = &self.region {
-                    iceberg_configs.insert(S3_REGION.to_string(), region.clone().to_string());
-                }
-                if let Some(endpoint) = &self.endpoint {
-                    iceberg_configs.insert(S3_ENDPOINT.to_string(), endpoint.clone().to_string());
-                }
-                iceberg_configs.insert(
-                    S3_ACCESS_KEY_ID.to_string(),
-                    self.access_key.clone().to_string(),
-                );
-                iceberg_configs.insert(
-                    S3_SECRET_ACCESS_KEY.to_string(),
-                    self.secret_key.clone().to_string(),
-                );
-                let config_builder = iceberg_catalog_glue::GlueCatalogConfig::builder()
-                    .warehouse(self.path.clone())
-                    .props(iceberg_configs);
-                let config = if let Some(uri) = self.uri.as_deref() {
-                    config_builder.uri(uri.to_string()).build()
-                } else {
-                    config_builder.build()
-                };
-                let catalog = iceberg_catalog_glue::GlueCatalog::new(config).await?;
-                Ok(Arc::new(catalog))
-            }
-            catalog_type if catalog_type == "hive" || catalog_type == "jdbc" => {
-                // Create java catalog
-                let (base_catalog_config, java_catalog_props) = self.build_jni_catalog_configs()?;
-                let catalog_impl = match catalog_type {
-                    "hive" => "org.apache.iceberg.hive.HiveCatalog",
-                    "jdbc" => "org.apache.iceberg.jdbc.JdbcCatalog",
-                    _ => unreachable!(),
-                };
-
-                jni_catalog::JniCatalog::build_catalog_v2(
-                    base_catalog_config,
-                    self.catalog_name(),
-                    catalog_impl,
-                    java_catalog_props,
-                )
-            }
-            _ => {
-                bail!(
-                    "Unsupported catalog type: {}, only support `storage`, `rest`, `hive`, `jdbc`, `glue`",
-                    self.catalog_type()
-                )
-            }
-        }
-    }
-
-    pub async fn load_table_v2(&self) -> ConnectorResult<TableV2> {
-        let catalog = self
-            .create_catalog_v2()
-            .await
-            .context("Unable to load iceberg catalog")?;
-
-        let table_id = self
-            .full_table_name_v2()
-            .context("Unable to parse table name")?;
-
-        catalog.load_table(&table_id).await.map_err(Into::into)
-    }
-
-    pub async fn load_table_v2_with_metadata(
-        &self,
-        metadata: TableMetadata,
-    ) -> ConnectorResult<TableV2> {
-        match self.catalog_type() {
-            "storage" => {
-                let config = StorageCatalogConfig::builder()
-                    .warehouse(self.path.clone())
-                    .access_key(self.access_key.clone())
-                    .secret_key(self.secret_key.clone())
-                    .region(self.region.clone())
-                    .endpoint(self.endpoint.clone())
-                    .build();
-                let storage_catalog = storage_catalog::StorageCatalog::new(config)?;
-
-                let table_id = self
-                    .full_table_name_v2()
-                    .context("Unable to parse table name")?;
-
-                Ok(iceberg::table::Table::builder()
-                    .metadata(metadata)
-                    .identifier(table_id)
-                    .file_io(storage_catalog.file_io().clone())
-                    // Only support readonly table for storage catalog now.
-                    .readonly(true)
-                    .build()?)
-            }
-            _ => self.load_table_v2().await,
-        }
     }
 }
 
@@ -708,7 +198,11 @@ impl IcebergSink {
 
         let table = self
             .config
-            .load_table()
+            .common
+            .load_table(
+                &self.config.path_style_access,
+                &self.config.java_catalog_props,
+            )
             .await
             .map_err(|err| SinkError::Iceberg(anyhow!(err)))?;
 
@@ -728,9 +222,17 @@ impl IcebergSink {
     }
 
     async fn create_table_if_not_exists(&self) -> Result<()> {
-        let catalog = self.config.create_catalog_v2().await?;
+        let catalog = self
+            .config
+            .common
+            .create_catalog_v2(
+                &self.config.path_style_access,
+                &self.config.java_catalog_props,
+            )
+            .await?;
         let table_id = self
             .config
+            .common
             .full_table_name_v2()
             .context("Unable to parse table name")?;
         if !catalog
@@ -738,7 +240,7 @@ impl IcebergSink {
             .await
             .map_err(|e| SinkError::Iceberg(anyhow!(e)))?
         {
-            let namespace = if let Some(database_name) = &self.config.database_name {
+            let namespace = if let Some(database_name) = &self.config.common.database_name {
                 NamespaceIdent::new(database_name.clone())
             } else {
                 bail!("database name must be set if you want to create table")
@@ -767,16 +269,16 @@ impl IcebergSink {
 
             let location = {
                 let mut names = namespace.clone().inner();
-                names.push(self.config.table_name.to_string());
-                if self.config.path.ends_with('/') {
-                    format!("{}{}", self.config.path, names.join("/"))
+                names.push(self.config.common.table_name.to_string());
+                if self.config.common.warehouse_path.ends_with('/') {
+                    format!("{}{}", self.config.common.warehouse_path, names.join("/"))
                 } else {
-                    format!("{}/{}", self.config.path, names.join("/"))
+                    format!("{}/{}", self.config.common.warehouse_path, names.join("/"))
                 }
             };
 
             let table_creation = TableCreation::builder()
-                .name(self.config.table_name.clone())
+                .name(self.config.common.table_name.clone())
                 .schema(iceberg_schema)
                 .location(location)
                 .build();
@@ -831,7 +333,7 @@ impl Sink for IcebergSink {
     const SINK_NAME: &'static str = ICEBERG_SINK;
 
     async fn validate(&self) -> Result<()> {
-        if "glue".eq_ignore_ascii_case(self.config.catalog_type()) {
+        if "glue".eq_ignore_ascii_case(self.config.common.catalog_type()) {
             risingwave_common::license::Feature::IcebergSinkWithGlue
                 .check_available()
                 .map_err(|e| anyhow::anyhow!(e))?;
@@ -876,7 +378,14 @@ impl Sink for IcebergSink {
     }
 
     async fn new_coordinator(&self) -> Result<Self::Coordinator> {
-        let catalog = self.config.create_catalog().await?;
+        let catalog = self
+            .config
+            .common
+            .create_catalog(
+                &self.config.path_style_access,
+                &self.config.java_catalog_props,
+            )
+            .await?;
         let table = self.create_and_validate_table().await?;
         let partition_type = table.current_partition_type()?;
 
@@ -1361,6 +870,7 @@ mod test {
 
     use risingwave_common::catalog::Field;
 
+    use crate::connector_common::IcebergCommon;
     use crate::sink::decouple_checkpoint_log_sink::DEFAULT_COMMIT_CHECKPOINT_INTERVAL_WITH_SINK_DECOUPLE;
     use crate::sink::iceberg::IcebergConfig;
     use crate::source::DataType;
@@ -1425,19 +935,20 @@ mod test {
         let iceberg_config = IcebergConfig::from_btreemap(values).unwrap();
 
         let expected_iceberg_config = IcebergConfig {
-            connector: "iceberg".to_string(),
+            common: IcebergCommon {
+                warehouse_path: "s3://iceberg".to_string(),
+                catalog_uri: Some("jdbc://postgresql://postgres:5432/iceberg".to_string()),
+                region: Some("us-east-1".to_string()),
+                endpoint: Some("http://127.0.0.1:9301".to_string()),
+                access_key: "hummockadmin".to_string(),
+                secret_key: "hummockadmin".to_string(),
+                catalog_type: Some("jdbc".to_string()),
+                catalog_name: Some("demo".to_string()),
+                database_name: Some("demo_db".to_string()),
+                table_name: "demo_table".to_string(),
+            },
             r#type: "upsert".to_string(),
             force_append_only: false,
-            table_name: "demo_table".to_string(),
-            database_name: Some("demo_db".to_string()),
-            catalog_name: Some("demo".to_string()),
-            catalog_type: Some("jdbc".to_string()),
-            path: "s3://iceberg".to_string(),
-            uri: Some("jdbc://postgresql://postgres:5432/iceberg".to_string()),
-            region: Some("us-east-1".to_string()),
-            endpoint: Some("http://127.0.0.1:9301".to_string()),
-            access_key: "hummockadmin".to_string(),
-            secret_key: "hummockadmin".to_string(),
             path_style_access: Some(true),
             primary_key: Some(vec!["v1".to_string()]),
             java_catalog_props: [("jdbc.user", "admin"), ("jdbc.password", "123456")]
@@ -1451,7 +962,7 @@ mod test {
         assert_eq!(iceberg_config, expected_iceberg_config);
 
         assert_eq!(
-            &iceberg_config.full_table_name().unwrap().to_string(),
+            &iceberg_config.common.full_table_name().unwrap().to_string(),
             "demo_db.demo_table"
         );
     }
@@ -1459,7 +970,14 @@ mod test {
     async fn test_create_catalog(configs: BTreeMap<String, String>) {
         let iceberg_config = IcebergConfig::from_btreemap(configs).unwrap();
 
-        let table = iceberg_config.load_table().await.unwrap();
+        let table = iceberg_config
+            .common
+            .load_table(
+                &iceberg_config.path_style_access,
+                &iceberg_config.java_catalog_props,
+            )
+            .await
+            .unwrap();
 
         println!("{:?}", table.table_name());
     }
