+++ src/connector/src/with_options.rs
@@ -66,7 +66,7 @@ impl WithOptions for i32 {}
 impl WithOptions for i64 {}
 impl WithOptions for f64 {}
 impl WithOptions for std::time::Duration {}
-impl WithOptions for crate::connector_common::mqtt_common::QualityOfService {}
+impl WithOptions for crate::connector_common::MqttQualityOfService {}
 impl WithOptions for crate::sink::kafka::CompressionCodec {}
 impl WithOptions for crate::source::filesystem::file_common::CompressionFormat {}
 impl WithOptions for nexmark::config::RateShape {}
