+++ src/connector/with_options_source.yaml
@@ -98,18 +98,24 @@ IcebergProperties:
     required: true
   - name: warehouse.path
     field_type: String
+    comments: Path of iceberg warehouse, only applicable in storage catalog.
     required: true
   - name: catalog.name
     field_type: String
+    comments: |-
+      Catalog name, can be omitted for storage catalog, but
+      must be set for other catalogs.
     required: false
   - name: catalog.uri
     field_type: String
+    comments: URI of iceberg catalog, only applicable in rest catalog.
     required: false
   - name: database.name
     field_type: String
     required: false
   - name: table.name
     field_type: String
+    comments: Full name of table, must include schema name.
     required: true
   - name: catalog.jdbc.user
     field_type: String
